// frontend/src/app/models/auth.models.ts
export interface OAuthAccount {
  id: string;
  oauth_name: string;
  account_id: string;
  account_email: string;
}
// ...
export interface UserRead {
  id: string; // UUID est une chaîne de caractères en TS/JS
  email: string;
  is_active: boolean;
  is_superuser: boolean;
  is_verified: boolean;
  pseudo?: string;
  picture?: string;
  given_name?: string;
  family_name?: string;
  locale?: string;
  oauth_accounts?: OAuthAccount[];
  education_level?: string;
  age?: number;
  ai_familiarity?: number;
  credits?: number;
  date_claim?: string;  // ISO date string
  role?: string; // Rôle de l'utilisateur (admin, contributor, user)
}